exploratory analysis
two.Rmd contains getting, cleaning, tokenization, and resulting dictionaries.  It displays bigrams, trigrams and unigrams top picks for most frequently encountered within the data frames described in text of two.Rmd. Training and testing sets are generated, original data is sampled. Tables and plots render for visualization of the data: wordclouds and bargraphs. The Appendix includes a study on models for determining accuracy statistics for comparison.